\name{Undocumented functions}
\alias{a2combination}
\alias{calculate_global_PSG}
\alias{calculate_indices}
\alias{calculate_singleton}
\alias{is_partition_group}
\alias{p_n_k}
\alias{p_tilde_n_k}
\alias{calculate_individual_PSG_approximative}
\alias{calculate_individual_PSG_exact}
\alias{calculate_individual_PSG}
\title{Undocumented functions}
\description{
  The following functions are for internal computation only: \emph{calculate_global_PSG, calculate_indices, calculate_singleton, is_partition_group, 
  p_n_k, p_tilde_n_k, calculate_individual_PSG_approximative, calculate_individual_PSG_exact, calculate_individual_PSG}.
}